Gut existing custom block system
